requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.13.3
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cymem==2.0.11
debugpy==1.8.13
decorator==5.2.1
deep-translator==1.11.4
dj-database-url==2.3.0
Django==5.1.7
django-cleanup==9.0.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
executing==2.2.0
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.12
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
preshed==3.0.9
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.3.0
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==77.0.3
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
thinc==8.3.4
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
whitenoise==6.9.0
wrapt==1.17.2
